requirements.txt
phidata
openai
python-dotenv
yfinance
requests